requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
celery
redis
pydantic[email]
alembic
python-dotenv
pydantic-settings